change: Move CCS stuff to it's own module
